http error response template 삭제 및 이름 변경
